+++ test/support/models.rb
@@ -214,3 +214,6 @@ class HashBackedAuthor < Hash
     'hash backed author'
   end
 end
+
+class UserNumber1And2 < User
+end